+++ interpreter/SyntacticAnalyzerStatus.java
@@ -6,4 +6,5 @@ package interpreter;
  */
 public class SyntacticAnalyzerStatus {
     public static final int SYNTAX_ERROR = -10;
+    public static final int ILLEGAL_TERMINATION = -11;
 }