+++ modeling/parser/expressions/AssignmentExpression.java
@@ -4,10 +4,12 @@ import modeling.Token;
 
 public class AssignmentExpression implements Expression {
     private final Token name;
+    private final Token operator;
     private final Expression value;
 
-    public AssignmentExpression(Token name, Expression value) {
+    public AssignmentExpression(Token name, Token operator, Expression value) {
         this.name = name;
+        this.operator = operator;
         this.value = value;
     }
 }