+++ modeling/parser/statements/FunctionStatement.java
@@ -4,11 +4,11 @@ import java.util.List;
 import modeling.Token;
 
 public class FunctionStatement implements Statement {
-    final Token name;
-    final List<Token> params;
-    final StatementBlock body;
+    private final Token name;
+    private final List<Token> params;
+    private final StatementBlock body;
 
-    FunctionStatement(Token name, List<Token> params, StatementBlock body) {
+    public FunctionStatement(Token name, List<Token> params, StatementBlock body) {
         this.name = name;
         this.params = params;
         this.body = body;